+++ packages/core/src/reducer.ts
@@ -8,7 +8,7 @@ import readyState from './reducers/readyState';
 import referenceGrammarID from './reducers/referenceGrammarID';
 import sendBoxValue from './reducers/sendBoxValue';
 import sendTimeout from './reducers/sendTimeout';
-import sendTyping from './reducers/sendTyping';
+import sendTypingIndicator from './reducers/sendTypingIndicator';
 import suggestedActions from './reducers/suggestedActions';
 
 export default combineReducers({
@@ -20,6 +20,9 @@ export default combineReducers({
   referenceGrammarID,
   sendBoxValue,
   sendTimeout,
-  sendTyping,
-  suggestedActions
+  sendTypingIndicator,
+  suggestedActions,
+
+  // TODO: [P3] Take this deprecation code out when releasing on or after January 13 2020
+  sendTyping: sendTypingIndicator
 })